+++ srcs/Conf/utils.cpp
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   utils.cpp                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: garra <garra@student.42.fr>                +#+  +:+       +#+        */
+/*   By: sahafid <sahafid@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/19 13:48:22 by sahafid           #+#    #+#             */
-/*   Updated: 2023/01/30 16:39:12 by garra            ###   ########.fr       */
+/*   Updated: 2023/01/31 12:06:35 by sahafid          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_stoi(std::string str)
 {
-	int	i;
+	size_t	i;
 	int	res;
 	int	sign;
 
@@ -56,7 +56,7 @@ std::vector<std::string> split(std::string s, char c)
     int del=0,i=0,p=0,j=0;
     
 	str = trim(initial, c);
-    for (int i = 0; i < str.length(); i++)
+    for (size_t i = 0; i < str.length(); i++)
 		if (str[i] == c && str[i + 1] != c)
 			del++;
     while (del-- >= 0)
